Move monitoring out of internal
This monitoring code is required to be called by users of the library and thus can't live in the internal directory.
