+++ internal/distribute/rest/distribute.go
@@ -28,7 +28,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/transparency-dev/formats/log"
 	"github.com/transparency-dev/witness/internal/config"
-	"github.com/transparency-dev/witness/internal/monitoring"
+	"github.com/transparency-dev/witness/monitoring"
 	"golang.org/x/mod/sumdb/note"
 )
